# orders-api/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Orders API
  version: "1.0.0"
  description: |
    REST API for managing customer orders and tracking information.
    Includes endpoints to list orders, fetch order details, get tracking info,
    and cancel orders or individual order items.

servers:
  - url: http://localhost:4000

paths:
  /orders:
    get:
      summary: Get all orders
      operationId: getAllOrders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "404":
          description: No orders found

  /orders/{id}:
    get:
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /orders/{id}/tracking:
    get:
      summary: Get tracking info for an order
      operationId: getTracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
        "404":
          description: Tracking not found

  /orders/{id}/cancel:
    post:
      summary: Cancel an entire order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrderResponse"
        "403":
          description: Order no longer eligible for cancellation
        "404":
          description: Order not found
        "409":
          description: Order already finalized (delivered or cancelled)

  /orders/{orderId}/items/{sku}/cancel:
    post:
      summary: Cancel a single item in an order
      operationId: cancelOrderItem
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelItemResponse"
        "403":
          description: Item no longer eligible for cancellation
        "404":
          description: Order or item not found
        "409":
          description: Item already finalized (delivered or cancelled)

components:
  schemas:
    Order:
      type: object
      properties:
        order_id:
          type: string
        placed_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [processing, in_transit, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderItem:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        qty:
          type: integer
        status:
          type: string
          enum: [processing, in_transit, delivered, cancelled]

    Tracking:
      type: object
      properties:
        order_id:
          type: string
        current_status:
          type: string
        eta:
          type: string
          format: date-time
        last_checkpoint:
          type: string

    CancelOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        cancelled:
          type: boolean
        error:
          type: string
          nullable: true

    CancelItemResponse:
      type: object
      properties:
        order_id:
          type: string
        sku:
          type: string
        cancelled:
          type: boolean
        error:
          type: string
          nullable: true